infra: avoid waiting for all instances to be running when empty 'infraEc2instance'
Previously: when 'infra' didn't describe any EC2 instances, provisioning
failed whenever any of our instances remained in a 'stopped' state for
the whole of 'DescribeInstances' waiter's polling period. This is
because when 'DescribeInstances' is passed an empty list of instance
IDs, it interprets this as a request for all of them.

Internally, we think of an empty 'ec2-instance' to mean we have no
expectations about EC2 instances, so just elide that 'await' in that
case.