Introduce update semantics based on graph traversal.
Resources are rendered into a DAG based on 'InfraRef', and are
registered as requiring update when any of their inputs change.

At this time, all updates are destructive; there are no resources for
which all attributes are mutable by 'Update*' actions, so for laziness's
sake whenever there's a change we're replacing them instead. I intend to
smarten up this logic for (at least) security groups and load balancers
soon.

This is a breaking change: absent hash annotations, any existing
deployments will be registered as requiring update, and our naive update
semantics make this destructive.