Forcibly detach EBS volumes when updating instances, and clean detach fails.
Usually this is because the volume hasn't been unmounted on the running
instance, leaving the attachment locked in a 'busy' state. It's usually
cleaner to unmount first, but that's somewhat out of scope.

The Amazonka bump is for two purposes: we need clean parses of the
VolumeAttachmentState 'busy' (12485eb), and rely on indication of
the terminal behaviour of waiters (4fece6f).